优化代码，调整变量、方法与类名，完善测试用例
