+++ hutool-core/src/main/java/cn/hutool/core/annotation/SynthesizedAnnotation.java
@@ -3,15 +3,15 @@ package cn.hutool.core.annotation;
 import cn.hutool.core.collection.CollUtil;
 
 import java.lang.annotation.Annotation;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.function.UnaryOperator;
 
 /**
  * <p>用于在{@link SynthesizedAnnotationAggregator}中表示一个处于合成状态的注解对象。<br>
- * 当对多个合成注解排序时，默认先按{@link #getVerticalDistance()}排序，
- * 再按{@link #getHorizontalDistance()}排序。
- * 该顺序应当保证当合成注解与其他注解存在层级关系时，
- * 离根对象最接近的注解被排在靠前的位置。
+ * 当对多个合成注解排序时，默认使用{@link #DEFAULT_CHILD_PRIORITY_COMPARATOR}进行排序，
+ * 从保证合成注解按{@link #getVerticalDistance()}与{@link #getHorizontalDistance()}的返回值保持有序，
+ * 从而使得距离根元素更接近的注解对象在被处理是具有更高的优先级。
  *
  * @author huangchengxing
  * @see SynthesizedAnnotationAggregator
@@ -19,7 +19,16 @@ import java.util.function.UnaryOperator;
 public interface SynthesizedAnnotation extends Annotation, Comparable<SynthesizedAnnotation> {
 
 	/**
-	 * 获取所属的合成注解
+	 * {@link SynthesizedAnnotation}使用的默认的比较器，
+	 * 按照按{@link #getVerticalDistance()}和{@link #getHorizontalDistance()}的返回值进行排序。<br>
+	 * 一般情况下，排序越小的合成注解应当被优先处理。
+	 */
+	Comparator<SynthesizedAnnotation> DEFAULT_CHILD_PRIORITY_COMPARATOR = Comparator
+		.comparing(SynthesizedAnnotation::getVerticalDistance)
+		.thenComparing(SynthesizedAnnotation::getHorizontalDistance);
+
+	/**
+	 * 获取所属的合成注解聚合器
 	 *
 	 * @return 合成注解
 	 */
@@ -114,7 +123,7 @@ public interface SynthesizedAnnotation extends Annotation, Comparable<Synthesize
 	 */
 	@Override
 	default int compareTo(SynthesizedAnnotation o) {
-		return SyntheticAnnotationUtil.getChildPriorityAnnotationCompare().compare(this, o);
+		return DEFAULT_CHILD_PRIORITY_COMPARATOR.compare(this, o);
 	}
 
 }
